server/requirements.txt
flask==3.1.0
flask_cors==5.0.0
python-dotenv==1.0.1
flask-pymongo==3.0.1